NO-ISSUE: test: consolidate API tests into web/playwright/
Move API test infrastructure from web/e2e/ into web/playwright/ to
co-locate with existing UI tests and share utilities:

- Helpers (RawApiClient, auth, image-ops) → playwright/utils/
- Shared code (csrf, test-utils) → playwright/shared/
- API fixtures, specs, config → playwright/api/
- ApiClient now uses shared requestCsrfToken instead of inline copy
- Handle 404 from /api/v1/user/initialize when feature is disabled
- Delete web/e2e/ entirely
