--- web/e2e/api-tests/helpers/index.ts
@@ -1,15 +1,0 @@
-export {test, expect, uniqueName} from './fixtures';
-export {RawApiClient} from './api-client';
-export {
-  initializeSuperuser,
-  getAccessToken,
-  createOAuthToken,
-  getV2Token,
-} from './auth';
-export {
-  pushImage,
-  pushImageAll,
-  orasAttach,
-  isSkopeoAvailable,
-  isOrasAvailable,
-} from './image-ops';
