+++ web/package.json
@@ -53,7 +53,7 @@
     "test": "react-scripts test",
     "format": "prettier --config .prettierrc \"src/**/*.{ts,tsx}\" --write",
     "test:integration": "cypress run",
-    "test:api": "playwright test --config e2e/api-tests/playwright.api.config.ts",
+    "test:api": "playwright test --config playwright/api/playwright.api.config.ts",
     "test:e2e": "playwright test",
     "test:e2e:ui": "playwright test --ui",
     "test:e2e:debug": "playwright test --debug",
