+++ web/playwright/utils/api/index.ts
@@ -3,6 +3,13 @@
  */
 
 export {ApiClient} from './client';
+export {RawApiClient} from './raw-client';
+export {
+  initializeSuperuser,
+  getAccessToken,
+  createOAuthToken,
+  getV2Token,
+} from './auth';
 export type {
   CreateRobotResponse,
   CreateUserResponse,
