+++ web/playwright/utils/image-ops.ts
